Move crypto code to a subdirectory
