+++ llarp/crypto/crypto.cpp
@@ -1,4 +1,4 @@
-#include <crypto.hpp>
+#include <crypto/crypto.hpp>
 
 #include <util/buffer.hpp>
