+++ llarp/messages/transfer_traffic.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_MESSAGES_TRANSFER_TRAFFIC_HPP
 #define LLARP_MESSAGES_TRANSFER_TRAFFIC_HPP
 
-#include <crypto.hpp>
+#include <crypto/crypto.hpp>
 #include <encrypted.hpp>
 #include <routing/message.hpp>
