+++ llarp/path/path_types.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_PATH_TYPES_HPP
 #define LLARP_PATH_TYPES_HPP
 
-#include <crypto.hpp>
+#include <crypto/crypto.hpp>
 #include <util/aligned.hpp>
 
 namespace llarp
