+++ llarp/router_contact.hpp
@@ -3,8 +3,9 @@
 
 #include <address_info.hpp>
 #include <constants/version.hpp>
-#include <crypto.hpp>
+#include <crypto/crypto.hpp>
 #include <exit_info.hpp>
+#include <util/aligned.hpp>
 #include <util/bencode.hpp>
 
 #include <vector>
