+++ test/test_llarp_router.cpp
@@ -1,6 +1,5 @@
-
 #include <router.hpp>
-#include <crypto.hpp>
+#include <crypto/crypto.hpp>
 
 #include <functional>
 #include <random>
